% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getgobp.R
\name{getgobp.community}
\alias{getgobp.community}
\title{Greate a table to record results. Every gene w's community takes over a row.}
\usage{
getgobp.community(graph, z.matrix, k = 2, n.cores = 4, cutoff = 1,
  community.min = 5, term.limit = NA)
}
\arguments{
\item{graph}{The graph of gene network.}

\item{z.matrix}{A matrix representing gene Z. Row names are the gene id in gene network.}

\item{k}{An Integer giving the order of the network.}

\item{n.cores}{A Core number used for parallel computing.}

\item{cutoff}{A number used to find LA scouting gene z.}

\item{community.min}{An Integer confines the min gene numbers in community.}

\item{term.limit}{A parameter indicates there is no limit of content in a line of the table.}
}
\value{
A form contains id of gene w, GO info of gene w , xk.w.semantic.similarity, x.w.avg.distance.
}
\description{
\code{getgobp.community()} generates a table. It can be used to record the information of  gene x,  gostats of x,
gostats of genes within k steps of x, gene w, gostats of w, the similarity of gene w and genes within k
steps of gene x, the average distance between gene x and gene w.
A gene x may correspond with several w communities. Thus one community takes a row in the table.
}

