% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lascouting.R
\name{lascouting}
\alias{lascouting}
\title{Find the liquid association scouting gene}
\usage{
lascouting(network.graph, express.matrix, k = 2, n.cores = 4)
}
\arguments{
\item{network.graph}{An igraph object representing the gene network.}

\item{express.matrix}{A matrix represeting the express matrix for the genes in gene network.Row names are the gene id in gene network.}

\item{k}{Integer giving the order of the network.}

\item{n.cores}{core number used for parallel computing}
}
\value{
A logical matrix representing the LA-scouting genes for each gene. Rows represent the center gene id and columns represents the LA-scouting genes.
}
\description{
Find the liquid association scouting gene
}

