% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.with.community}
\alias{visualize.with.community}
\title{visualize with w community: Generate a graph of genes w and their community in different colors.}
\usage{
visualize.with.community(graph, kernel.result, x, k = 2, cutoff = 1,
  community.min = 5, path = NULL)
}
\arguments{
\item{graph}{The fraph of gene network.}

\item{kernel.result}{The result of kernel.density which finds genes z of a gene x.}

\item{x}{The Gene the graph is generated for.}

\item{k}{A specific number stands for the length of step from gene y to gene x.}

\item{cutoff}{A specific number to filter gene w from gene z.}

\item{path}{The path where the result graph is saved to.The default path is the original path of input graph.}

\item{cummunity.min}{An Integer confines the least number of genes in a community of w shown in graph.}
}
\value{
a graph displays genes w and their correspongding community in different colors.
}
\description{
visualize with w community: Generate a graph of genes w and their community in different colors.
}
\examples{
# Create sample data for examples.
graph <- erdos.renyi.game(50,0.3)
relate_matrix <- matrix(date=rexp(200,rate=.1), nrow=50, ncol=5, byrow= TURE, dimnames=NULL)
#use the first normalize method as an example
kernel.result <- kernel.density(relate_matrix, graph, smoothing.normalize=c("one"))
visualize.with.community(graph,kernel.result,x,k=2,cutoff=1,community.min=5,path=NULL)
}
\seealso{
\code{\link{kernel.result}}
}

