% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{xw.distance}
\alias{xw.distance}
\title{Create a table to record the distance between gene x and gene w.}
\usage{
xw.distance(graph, z.matrix, cutoff = 0.8, n.cores = 4)
}
\arguments{
\item{graph}{The graph of the gene network.}

\item{z.matrix}{A matrix representing gene Z. Row names are the gene id in gene network.}

\item{cutoff}{A number used to find LA scouting gene z.}

\item{n.cores}{Core number used for parallel computing.}
}
\value{
a table contains distance between all genes x and their correspongding genes w.
}
\description{
\code{xw.distance()} generates a table contains distance between all genes x and their correspongding genes w.
}
\examples{
# Create sample data for examples.
z.matrix <- matrix(data=rexp(200,rate=.1), nrow=50, ncol=5, byrow= TRUE, dimnames=NULL)
graph <- erdos.renyi.game(50,0.3)
xw.distance(graph, z.matrix, cutoff=0.8, n.cores=4)
}

