% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kernel.density.R
\name{kernel.density}
\alias{kernel.density}
\title{Evaluate the result using kernel density estimation.}
\usage{
kernel.density(relate.matrix, network.graph, smoothing.normalize = c("one",
  "squareM", "none"))
}
\arguments{
\item{relate.matrix}{The matrix returned by lascouting.}

\item{network.graph}{The igraph object representing the gene network.}

\item{smoothing.normalize}{Different ways to normalize the result.}
}
\value{
A  matrix representing the kernel density of each gene. Each row is a gene, columns are the weights of scouting genes for the gene.
}
\description{
Evaluate the result using kernel density estimation.
}

